requirements.txt
websockets>=15.0
httpx>=0.28
python-dotenv>=1.0
anthropic>=0.40
rich>=13.0
